+++ src/renderInput.js
@@ -3,6 +3,7 @@ import { PropTypes, createElement } from 'react'
 import { Input } from 'uikit-react'
 
 const renderInput = ({
+  autoComplete,
   input,
   meta: { asyncValidating, error, touched },
   label,
@@ -16,8 +17,11 @@ const renderInput = ({
   wrapperClassName: customWrapperClassName,
   ...custom
 }) => {
+  if (autoComplete === false) {
+    // console.error('autoComplete is no longer set automatically in redux-form-uikit. You should review code that rely on this being set automatically. autoComplete={false} can be safely removed, however.')
+  }
+
   const component = createElement(inputComponent, {
-    autoComplete: input.name,
     id: label ? input.name : undefined,
     placeholder: label,
     ...input,
@@ -76,12 +80,13 @@ renderInput.propTypes = {
   input: PropTypes.shape({
     name: PropTypes.string.isRequired,
   }).isRequired,
-  inputComponent: PropTypes.node.isRequired,
+  inputComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]).isRequired,
   meta: PropTypes.shape({
     asyncValidating: PropTypes.bool,
     error: PropTypes.string,
     touched: PropTypes.bool,
   }).isRequired,
+  autoComplete: PropTypes.string,
   errorClassName: PropTypes.string,
   errorDisplay: PropTypes.oneOf(['inline', 'block']),
   help: PropTypes.node,