+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Changelog.
 
+## Fixed
+* `inputComponent` prop did not allow strings.
+
+## Deprecated
+* `autoComplete` will no longer be added automatically in next major release.
+
 ### Documentation
 * using Yarn.